resolve conflicts.
