+++ src/redux/actions/missions.js
@@ -1,15 +1,8 @@
 import { FETCH_ALL_MISSIONS } from './actionTypes';
-import * as api from '../../api/api';
 
-// API action creators
-const getMissions = () => async (dispatch) => {
-  try {
-    const data = await api.fetchMissions();
+const addMission = (payload) => ({
+  type: FETCH_ALL_MISSIONS,
+  payload,
+});
 
-    dispatch({ type: FETCH_ALL_MISSIONS, payload: data });
-  } catch (error) {
-    throw new Error(error.message);
-  }
-};
-
-export default getMissions;
+export default addMission;
